Allow static/abstract classifier also for official UML syntax
Previously, the static classifier was only supported for attributes
using the syntax "String name$". Now you can also write "name: String$"
See PR #2269 and issue #3001 and my old issue #1953

Further, it was confusing that static/abstract classifiers for methods
had to be written after the closing parenthesis of the method signature,
even if the method has a return type. So you had to write
"foo(name: String)* int".

This commit also adds the possibility of writing the classifier at the
very end, as in "foo(name: String) int*".