Change parseMessage to parseText
This allows for a conceptual and type splitting between texts which
appear in many places in the diagram (messages, notes,
actor descriptions, etc) and messages between participants.

I believe it makes more sense to call parseText rather than
parseMessage when used in context of working with other components
in the diagram which will help with developer experience.

A notable change when implementing this is changing all wrap
fields in relevant interfaces to optional. This is done so that
if left undefined, we will resolt to an autoWrap function which
references the diagram config. This will allow for an implementation
of a diagram wide default of either wrap or nowrap that's left to the
user to decide, instead of a hard-coded default.