ci(release-drafter): remove unused `branch` config
`branch` is not a valid configuration option for release-drafter,
see
https://github.com/release-drafter/release-drafter#configuration-options

There's is a similar [`references` option][1], but it does nothing when
using GitHub Actions (it's only there for GitHub apps).

There's also `commitish`, but it defaults to the target/branch the
GitHub Action job runs on, so we don't need to set that.

[1]: https://github.com/release-drafter/release-drafter#references