Add LLM integration section in productivity tools and add HueHive
Since mermaid diagram creation is a perfact usecase for LLMs adding a section for LLM integration make sense. HueHive is a tool to create and visualize mermaid diagrams with the help of ChatGPT.
Please let me know if it is fine to add this or this should be added in some other section. Thanks.

PS: I am creater of HueHive diagrams tool.