chore: Update argos token
We have added a check in e2e, to ensure events in forks will not trigger the visual tests.

This will not protect against malicious users, but there is no other way to support PRs from forks.

A proposal has been sent to Argos team to add a secure approach for public repos, fingers crossed.