fix: fix direction parsing
The extraction regex ignored all characters before or after a direction,
which could hide syntax errors or remove parts of the graph.

Before correction, this line is a **valid** direction:
```mmd
bla bla direction RLblablabla
```

This is not a graph.
```mmd
direction LR@--> up
```

Neither that
```mmd
a --> b["direction LRTB"]
```