Fix C4 description text fitting
Fix C4 description text that overflowed the shape. The reason was that
the font size was never set in the `drawC4Shape` method and specifically
in the `_drawTextCandidateFunc` call, so it defaulted to the default
font size of 16px while calculating for 12px.

Update the description font size to not be the configured font size
minus 2 but equal to the configured font size so that it is better
readable but still smaller than the label.